Implement flight search
Create the form to accept from airport, to airport, passengers, and flight date from the user.
Search the flights table for all flights matching the criteria entered by the user.
Display the returned flights to the user on the same /flights page and allow the user to select a flight. (Note: the 'Choose Flight' submit current points to the same /flight page, but this will be changed.)